move dataclass local -> model.database
