+++ shared/src/commonMain/kotlin/com/kmm/moonflower/core/local/PlantDataSource.kt
@@ -1,5 +1,7 @@
 package com.kmm.moonflower.core.local
 
+import com.kmm.moonflower.core.model.database.Plant
+
 interface PlantDataSource {
 
     suspend fun insertOrReplacePlants(plants: List<Plant>)
