+++ shared/src/commonMain/kotlin/com/kmm/moonflower/core/local/PlantDataSourceImpl.kt
@@ -1,5 +1,6 @@
 package com.kmm.moonflower.core.local
 
+import com.kmm.moonflower.core.model.database.Plant
 import com.kmm.moonflower.database.AppDatabase
 import database.AppDatabaseQueries
 import database.Plants
